Clean up and modified and current time added
