+++ app/assets/stylesheets/components/_activity_field.scss
@@ -1,3 +1,7 @@
 #starting-time {
   visibility: hidden;
 }
+
+#current-time {
+  visibility: hidden;
+}
