+++ app/assets/stylesheets/components/_donut.scss
@@ -3,11 +3,3 @@
   width: 100px;
   height: 100px;
 }
-
-.time {
-      font-size: 12px;
-      color: black;
-      transform-origin: 100%;
-      transition: all 0.05s;
-      transition-timing-function: 1;
-    }
