+++ app/views/pages/profile.html.erb
@@ -34,7 +34,8 @@
     <p class="time-status"><%= time %></p>
   </div>
 
-  <p>Current time: <%= (Time.now.utc + 2.hours).strftime("%H:%M") %></p>
+
+  <p class="curr-time">Current time: <%= (Time.now.utc + 2.hours).strftime("%H:%M") %></p>
 
 <% else %>
 <!-- If the user is NOT charging... -->
